Rename `FilterablePokedex` -> `App`
