+++ src/index.js
@@ -3,14 +3,14 @@ import './style/styles.css'
 import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
-import FilterablePokedex from './containers/FilterablePokedex'
+import App from './containers/App'
 import configureStore from './redux/store/configureStore'
 
 const store = configureStore()
 
 render(
 	<Provider store={store}>
-		<FilterablePokedex/>
+		<App />
 	</Provider>,
 	document.getElementById('root')
 )
